# backend/internal/handler/openapi.yaml
openapi: 3.0.3
info:
  title: 影策 HTTP API
  version: "1.0.0"
  description: |
    路径前缀是 `/api`。前后端一起发布，不按路径做 API 版本。

    业务 JSON 信封为 `{ code, data, msg, reason }`。`code === 0` 表示成功；失败时 HTTP status 与
    `code` 都表达失败。`code` 默认等于 HTTP 状态；同一状态下的细分原因使用 `status*100+n`
    （例如 40301 配额、42901 限流）。机器判断请用 `reason`，不要解析 `msg`。

    TypeScript 调用面仍是 `web/src/services/api` 的 `http` 模块，不从本规范生成客户端。
servers:
  - url: /api
security:
  - sessionCookie: []
...
paths:
...
  /agent/workspaces/{canvasId}:
    parameters:
      - {in: path, name: canvasId, required: true, schema: {type: string}}
    get:
      tags: [agent]
      summary: 读取画布 Workspace（懒创建一对一记录）
      responses:
        "200": {description: 信封 data 为 WorkspaceView，含 workspaceId、canvasId、revision、agentsMd、agentsMdHash 和版本化 skills}
        "401": {description: 未登录}
        "404": {description: 画布不存在或不属于当前用户}
    patch:
      tags: [agent]
      summary: 使用 revision CAS 更新项目文档，仅影响新 Run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [revision, agentsMd]
              properties:
                revision: {type: integer, format: int64, minimum: 0}
                agentsMd: {type: string, description: UTF-8 文档最多 65536 字节；空字符串用于清空}
      responses:
        "200": {description: 更新后的 WorkspaceView}
        "400": {description: 非法 JSON、缺失字段或文档超限}
        "404": {description: 画布不存在或不属于当前用户}
        "409": {description: "reason=agent_workspace_revision_conflict，details.currentRevision 为当前 revision"}
  /agent/workspaces/{canvasId}/skills:
    parameters:
      - {in: path, name: canvasId, required: true, schema: {type: string}}
    get:
      tags: [agent]
      summary: 读取 Workspace 技能版本、来源与容量
      responses:
        "200": {description: 有序技能数组，含 skillId、skillVersionId、position、enabled、skillName、contentHash、fileCount、totalBytes、source=workspace}
        "404": {description: 画布不存在或不属于当前用户}
    put:
      tags: [agent]
      summary: CAS 替换 Workspace 技能追加项（清空传空数组）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [revision, skills]
              properties:
                revision: {type: integer, format: int64, minimum: 0}
                skills:
                  type: array
                  items:
                    type: object
                    additionalProperties: false
                    required: [skillId, skillVersionId, position, enabled]
                    properties:
                      skillId: {type: string}
                      skillVersionId: {type: string}
                      position: {type: integer, minimum: 0}
                      enabled: {type: boolean}
      responses:
        "200": {description: 更新后的 WorkspaceView，文档与技能共用 revision}
        "400": {description: 非法技能、私有技能、重复排序或容量超限}
        "404": {description: 画布不存在或不属于当前用户}
        "409": {description: "reason=agent_workspace_revision_conflict"}
...
  /tasks:
    post:
      tags: [tasks]
      summary: 创建生成任务
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "400":
          $ref: "#/components/responses/Error"
  /tasks/{id}/recover-media:
    post:
      tags: [tasks]
      summary: 按原任务 ID 重试作品保存，不重新生成或创建计费订单
      description: 仅当前用户有恢复记录的保存失败任务可重入队；已在队列或运行中时幂等返回。每任务最多手动恢复三次，间隔至少一分钟。已取消、已退款或归属项目不可用时拒绝。
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /tasks/{id}/text-events:
    get:
      tags: [tasks]
      summary: 文本任务 SSE
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: after
          schema: { type: integer }
          description: 递增事件游标，不要用任务 ID
      responses:
        "200":
          description: text/event-stream
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: open_ai_canvas_session
  schemas:
...
    Envelope:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
          description: 0 成功；失败时默认等于 HTTP 状态，或 status*100+n 的细分码
        data: {}
        msg:
          type: string
        reason:
          type: string
          description: 仅失败时出现。稳定机器可读原因。
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            code:
              type: integer
              enum: [400, 401, 403, 404, 409, 429, 500, 502, 503, 504, 40301, 40901, 42901]
            data:
              nullable: true
            reason:
              type: string
              enum:
                - invalid_argument
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - failed_precondition
                - quota_exceeded
                - rate_limited
                - unavailable
                - timeout
                - internal
                - bad_gateway
                - upstream_dns_failed
                - model_capability_not_supported
                - model_price_not_configured
                - model_route_unavailable
                - provider_request_failed
                - model_catalog_mismatch
                - invalid_model_selection
  responses:
...
    Envelope:
      description: 业务成功信封
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    Error:
      description: 业务失败信封
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"